[Patch] Added a check for names of choices in update_dropdown
In the current version, update_material_dropdown needs to be passed a named character vector to run. Without any names, the result is just a blank dropdown, which is confusing from the dev point of view.

As these name are mandatory for this to work, this little patch just add an extra check in case of missing names for the choices params.